update imports and cleanup remaining *.new forks
