--- packages/react-reconciler/src/ReactEventPriorities.new.js
@@ -1,82 +1,0 @@
-/**
- * Copyright (c) Meta Platforms, Inc. and affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- *
- * @flow
- */
-
-import type {Lane, Lanes} from './ReactFiberLane.new';
-
-import {
-  NoLane,
-  SyncLane,
-  InputContinuousLane,
-  DefaultLane,
-  IdleLane,
-  getHighestPriorityLane,
-  includesNonIdleWork,
-} from './ReactFiberLane.new';
-
-export opaque type EventPriority = Lane;
-
-export const DiscreteEventPriority: EventPriority = SyncLane;
-export const ContinuousEventPriority: EventPriority = InputContinuousLane;
-export const DefaultEventPriority: EventPriority = DefaultLane;
-export const IdleEventPriority: EventPriority = IdleLane;
-
-let currentUpdatePriority: EventPriority = NoLane;
-
-export function getCurrentUpdatePriority(): EventPriority {
-  return currentUpdatePriority;
-}
-
-export function setCurrentUpdatePriority(newPriority: EventPriority) {
-  currentUpdatePriority = newPriority;
-}
-
-export function runWithPriority<T>(priority: EventPriority, fn: () => T): T {
-  const previousPriority = currentUpdatePriority;
-  try {
-    currentUpdatePriority = priority;
-    return fn();
-  } finally {
-    currentUpdatePriority = previousPriority;
-  }
-}
-
-export function higherEventPriority(
-  a: EventPriority,
-  b: EventPriority,
-): EventPriority {
-  return a !== 0 && a < b ? a : b;
-}
-
-export function lowerEventPriority(
-  a: EventPriority,
-  b: EventPriority,
-): EventPriority {
-  return a === 0 || a > b ? a : b;
-}
-
-export function isHigherEventPriority(
-  a: EventPriority,
-  b: EventPriority,
-): boolean {
-  return a !== 0 && a < b;
-}
-
-export function lanesToEventPriority(lanes: Lanes): EventPriority {
-  const lane = getHighestPriorityLane(lanes);
-  if (!isHigherEventPriority(DiscreteEventPriority, lane)) {
-    return DiscreteEventPriority;
-  }
-  if (!isHigherEventPriority(ContinuousEventPriority, lane)) {
-    return ContinuousEventPriority;
-  }
-  if (includesNonIdleWork(lane)) {
-    return DefaultEventPriority;
-  }
-  return IdleEventPriority;
-}
