+++ packages/react-reconciler/src/ReactFiberCacheComponent.js
@@ -13,7 +13,7 @@ import type {Fiber} from 'react-reconciler/src/ReactInternalTypes';
 import {enableCache} from 'shared/ReactFeatureFlags';
 import {REACT_CONTEXT_TYPE} from 'shared/ReactSymbols';
 
-import {pushProvider, popProvider} from './ReactFiberNewContext.new';
+import {pushProvider, popProvider} from './ReactFiberNewContext';
 import * as Scheduler from 'scheduler';
 
 // In environments without AbortController (e.g. tests)
