+++ packages/react-reconciler/src/ReactFiberOffscreenComponent.js
@@ -8,13 +8,13 @@
  */
 
 import type {ReactNodeList, OffscreenMode, Wakeable} from 'shared/ReactTypes';
-import type {Lanes} from './ReactFiberLane.old';
-import type {SpawnedCachePool} from './ReactFiberCacheComponent.new';
+import type {Lanes} from './ReactFiberLane';
+import type {SpawnedCachePool} from './ReactFiberCacheComponent';
 import type {Fiber} from './ReactInternalTypes';
 import type {
   Transition,
   TracingMarkerInstance,
-} from './ReactFiberTracingMarkerComponent.new';
+} from './ReactFiberTracingMarkerComponent';
 
 export type OffscreenProps = {
   // TODO: Pick an API before exposing the Offscreen type. I've chosen an enum
