+++ packages/react-reconciler/src/ReactFiberRoot.js
@@ -14,18 +14,18 @@ import type {
   TransitionTracingCallbacks,
 } from './ReactInternalTypes';
 import type {RootTag} from './ReactRootTags';
-import type {Cache} from './ReactFiberCacheComponent.new';
+import type {Cache} from './ReactFiberCacheComponent';
 import type {Container} from './ReactFiberHostConfig';
 
 import {noTimeout, supportsHydration} from './ReactFiberHostConfig';
-import {createHostRootFiber} from './ReactFiber.new';
+import {createHostRootFiber} from './ReactFiber';
 import {
   NoLane,
   NoLanes,
   NoTimestamp,
   TotalLanes,
   createLaneMap,
-} from './ReactFiberLane.new';
+} from './ReactFiberLane';
 import {
   enableSuspenseCallback,
   enableCache,
@@ -34,9 +34,9 @@ import {
   enableUpdaterTracking,
   enableTransitionTracing,
 } from 'shared/ReactFeatureFlags';
-import {initializeUpdateQueue} from './ReactFiberClassUpdateQueue.new';
+import {initializeUpdateQueue} from './ReactFiberClassUpdateQueue';
 import {LegacyRoot, ConcurrentRoot} from './ReactRootTags';
-import {createCache, retainCache} from './ReactFiberCacheComponent.new';
+import {createCache, retainCache} from './ReactFiberCacheComponent';
 
 export type RootState = {
   element: any,
