+++ packages/react-reconciler/src/ReactFiberTracingMarkerComponent.js
@@ -13,11 +13,11 @@ import type {
   FiberRoot,
 } from './ReactInternalTypes';
 import type {OffscreenInstance} from './ReactFiberOffscreenComponent';
-import type {StackCursor} from './ReactFiberStack.new';
+import type {StackCursor} from './ReactFiberStack';
 
 import {enableTransitionTracing} from 'shared/ReactFeatureFlags';
-import {createCursor, push, pop} from './ReactFiberStack.new';
-import {getWorkInProgressTransitions} from './ReactFiberWorkLoop.new';
+import {createCursor, push, pop} from './ReactFiberStack';
+import {getWorkInProgressTransitions} from './ReactFiberWorkLoop';
 
 export type SuspenseInfo = {name: string | null};
